update zIndex type

+++ Unit Tests/InitialShotsCollectionViewLayoutSpec.swift
@@ -201,7 +201,7 @@ private struct FixtureFirstItemAttributes {
     let width = CGFloat(265)
     let height = CGFloat(198.75)
     let center = CGPoint(x: 160, y: 284)
-    let zIndex = NSInteger(0)
+    let zIndex = 0
     let alpha = CGFloat(0)
 }
 
@@ -209,6 +209,6 @@ private struct FixtureSecondItemAttributes  {
     let width = CGFloat(210)
     let height = CGFloat(157.5)
     let center = CGPoint(x: 160, y: 324)
-    let zIndex = NSInteger(-1)
+    let zIndex = -1
     let alpha = CGFloat(0)
 }
